Move License_Manager under License namespace + update providers
