+++ src/Uplink/Uplink.php
@@ -36,16 +36,16 @@ class Uplink {
 		$container->singleton( Site\Data::class, Site\Data::class );
 		$container->singleton( Notice\Provider::class, Notice\Provider::class );
 		$container->singleton( Admin\Provider::class, Admin\Provider::class );
-		$container->singleton( Auth\Provider::class, Auth\Provider::class );
 		$container->singleton( License\Provider::class, License\Provider::class );
+		$container->singleton( Auth\Provider::class, Auth\Provider::class );
 
 		if ( static::is_enabled() ) {
 			$container->get( View\Provider::class )->register();
 			$container->get( API\V3\Provider::class )->register();
 			$container->get( Notice\Provider::class )->register();
 			$container->get( Admin\Provider::class )->register();
-			$container->get( Auth\Provider::class )->register();
 			$container->get( License\Provider::class )->register();
+			$container->get( Auth\Provider::class )->register();
 		}
 
 		require_once __DIR__ . '/functions.php';
