+++ src/Uplink/functions.php
@@ -10,10 +10,10 @@ use StellarWP\Uplink\API\Validation_Response;
 use StellarWP\Uplink\Auth\Admin\Disconnect_Controller;
 use StellarWP\Uplink\Auth\Auth_Url_Builder;
 use StellarWP\Uplink\Auth\Authorizer;
-use StellarWP\Uplink\Auth\License\License_Manager;
 use StellarWP\Uplink\Auth\Token\Token_Factory;
 use StellarWP\Uplink\Components\Admin\Authorize_Button_Controller;
 use StellarWP\Uplink\License\License_Key_Fetcher;
+use StellarWP\Uplink\License\Manager\License_Manager;
 use StellarWP\Uplink\License\Storage\License_File_Storage;
 use StellarWP\Uplink\License\Storage\License_Network_Storage;
 use StellarWP\Uplink\License\Storage\License_Single_Site_Storage;
